Upgraded following packages xserver          commit b306df5a6060beea82b5157c3603593527b220b0 libxcb/xcb-proto commit 4d2879ad9e394ff832762e8961eca9415cc9934c xkeyboard-config commit dd0d8f7f06406628f43e4f06ceb7370fbd06c237 libX11           commit f2ebbce6d0978d2d84f840196b03ee35a3ca1736 libXau           commit 6b2e9a63d8306282a2b384762bea004c19f301a1 xkbcomp          commit 3f354d20f5bc0276e27faa7fb5f6356319aa32c3 pixman           commit 11e51bc72f37812674d49292a41a0b9721eb2077 libxtrans        commit ae99ac32f61e0db92a45179579030a23fe1b5770 fontconfig       commit 6d0580b2427f50b45e1432ba0fb6128c19b6dbd4 mesa             commit 019770f026ad1e1a9b03ddacab2df252126fd636 putty            commit bed0a8b596f156f2eb86d8e8e92eac31cc08fbfa xorgproto        commit af7cb6a643db810536605feab1402654a9818569

+++ mesalib/src/gallium/drivers/vc4/vc4_uniforms.c
@@ -316,7 +316,7 @@ vc4_write_uniforms(struct vc4_context *vc4, struct vc4_compiled_shader *shader,
                                 if (format_swiz[i] >= 4)
                                         continue;
 
-                                color |= (vc4->blend_color.ub[format_swiz[i]] <<
+                                color |= ((uint32_t)vc4->blend_color.ub[format_swiz[i]] <<
                                           (i * 8));
                         }
                         cl_aligned_u32(&uniforms, color);
@@ -324,7 +324,7 @@ vc4_write_uniforms(struct vc4_context *vc4, struct vc4_compiled_shader *shader,
                 }
 
                 case QUNIFORM_BLEND_CONST_COLOR_AAAA: {
-                        uint8_t a = vc4->blend_color.ub[3];
+                        uint32_t a = vc4->blend_color.ub[3];
                         cl_aligned_u32(&uniforms, ((a) |
                                                    (a << 8) |
                                                    (a << 16) |
